# product-api/handlers/products.go
package handlers

import (
	"log"
	"net/http"
	"strconv"

	protos "github.com/fzft/go_microservice/currency/protos/currency"
	"github.com/gorilla/mux"
)

// GenericError is a generic error message returned by a server
type GenericError struct {
	Message string `json:"message"`
}

// getProductID returns the product ID from the URL
// Panics if cannot convert the id into an integer
// this should never happen as the router ensures that
// this is a valid number
func getProductID(r *http.Request) int {
	// parse the product id from the url
	vars := mux.Vars(r)

	// convert the id into an integer and return
	id, err := strconv.Atoi(vars["id"])
	if err != nil {
		// should never happen
		panic(err)
	}

	return id
}

// Products handler for getting and updating products
type Products struct {
	l *log.Logger
	cc protos.CurrencyClient
}

// ValidationError is a collection of validation error messages
type ValidationError struct {
	Messages []string `json:"messages"`
}

// NewProducts returns a new products handler with the given logger
func NewProducts(l *log.Logger, cc protos.CurrencyClient) *Products {
	return &Products{l, cc}
}

// KeyProduct is a key used for the Product object in the context
type KeyProduct struct {
}